// packages/wp-react-lib/src/types.ts
export interface DgSettings {
    react_ui_url:          string;
    react_api_url:         string | null;
    apache_superset_url:   string | boolean | null;
    react_search_type:     string;
    react_menu_type:       string;
    languages:             Languages;
    landing_page_url:      string;
    google_analytics_code: string;
    name:                  string;
    description:           string;
    site_logo:             number;
    site_icon:             number;
}

export interface Languages {
    [key: string]: Language;
}

export interface Language {
    enable:          number;
    locale:          string;
    name:            string;
    translation:     string;
    date:            string;
    time:            string;
    flag:            string;
    wpseo_og_locale: string;
}

export interface Taxonomy {
    id:             number;
    name:           string;
    slug:           string;
    description:    string;
    types:          string[];
    hierarchical:   boolean;
    rest_base:      string;
    rest_namespace: string;
    _links:         Links;
    taxonomy:       string;
}

export interface Links {
    collection: Collection[];
    "wp:items": Collection[];
    curies:     Cury[];
}

export interface Collection {
    href: string;
}

export interface Cury {
    name:      string;
    href:      string;
    templated: boolean;
}
// ...
export interface WpCategory {
    id:          number;
    count:       number;
    description: string;
    link:        string;
    name:        string;
    slug:        string;
    taxonomy:    string;
    parent:      number;
    meta:        Record<string, unknown>;
    _links:      Links;
}

export interface ApiError {
    status:     number;
    statusText: string;
    url:        string;
}

export interface ApiResponse<T> {
    data: T;
    /** HTTP response headers (e.g., x-wp-total, x-wp-totalpages). */
    meta: Record<string, string>;
}

export interface Media {
    id:             number;
    date:           string;
    date_gmt:       string;
    guid:           Caption;
    modified:       string;
    modified_gmt:   string;
    slug:           string;
    status:         Status;
    type:           Type;
    link:           string;
    title:          Caption;
    author:         number;
    featured_media: number;
    comment_status: CommentStatus;
    ping_status:    PingStatus;
    template:       string;
    meta:           Meta;
    bread_crumbs:   any[];
    class_list:     string[];
    acf:            any[];
    meta_fields:    MetaFields;
    description:    Caption;
    caption:        Caption;
    alt_text:       string;
    media_type:     MediaType;
    mime_type:      MIMEType;
    media_details:  MediaDetails;
    post:           null;
    source_url:     string;
    _links:         Links;
}
// ...
export interface Caption {
    rendered: string;
}


export enum CommentStatus {
    Open = "open",
}

export interface MediaDetails {
    filesize: number;
    sizes:    Sizes;
}

export interface Sizes {
}

export enum MediaType {
    File = "file",
}

export interface Meta {
    _acf_changed:             boolean;
    inline_featured_image:    boolean;
    redirect_url:             string;
    myguten_meta_block_field: string;
}

export interface MetaFields {
    _wp_attached_file:       string[];
    _wp_attachment_metadata: string[];
    _edit_lock?:             string[];
}

export enum MIMEType {
    ApplicationJSON = "application/json",
}

export enum PingStatus {
    Closed = "closed",
}

export enum Status {
    Inherit = "inherit",
}

export enum Type {
    Attachment = "attachment",
}

export interface Menu {
    term_id:          number;
    name:             string;
    slug:             string;
    term_group:       number;
    term_taxonomy_id: number;
    taxonomy:         string;
    description:      string;
    parent:           number;
    count:            number;
    filter:           string;
    term_order:       string;
    icon_media_id?:    number;
}

export interface SearchResult {
    id:      number;
    title:   string;
    url:     string;
    type:    string;
    subtype: string;
    _links:  Links;
    value:   string;
}

export type SearchResults = SearchResult[];